refactor: BaseLayout 컴포넌트로 리팩터링
