+++ packages/web-domains/src/result/features/common/components/HeaderBanner/styles.ts
