+++ packages/web-domains/src/result/features/common/components/index.ts
@@ -1,1 +1,2 @@
 export { Section } from './Section/Section';
+export { BaseLayout } from './Layout/BaseLayout';
