+++ packages/web-domains/src/result/features/main/components/index.ts
@@ -1,4 +1,3 @@
-export { HeaderBanner } from './HeaderBanner/HeaderBanner';
 export { CountByMemberList } from './SelectedDataCard/CountByMemberList';
 export { MostAnswered, WithMeMembers } from './SelectedDataCard/Content';
 export { Comment } from './SelectedDataCard/Comment';
