+++ packages/web-domains/src/result/screens/ResultMainScreen.tsx
@@ -1,24 +1,12 @@
-import { colors, size } from '@sambad/sds/theme';
-import { HeaderBanner } from '../features/main/components';
 import { MostAnsweredContainers, WithMyMembersContainers, CommentListContainer } from '../features/main/containers';
+import { BaseLayout } from '../features/common/components';
 
 export const ResultMainScreen = () => {
   return (
-    <div style={layoutStyle}>
-      <HeaderBanner title="즐겨하는 스포츠는?" thumbnail="" />
-      <div style={contentStyle}>
-        <MostAnsweredContainers />
-        <WithMyMembersContainers />
-        <CommentListContainer />
-      </div>
-    </div>
+    <BaseLayout>
+      <MostAnsweredContainers />
+      <WithMyMembersContainers />
+      <CommentListContainer />
+    </BaseLayout>
   );
-};
-
-const layoutStyle = {
-  backgroundColor: colors.grey200,
-};
-
-const contentStyle = {
-  padding: size.xs,
 };
